Cooperate with linkcheck builder
Redirect links that direct to an HTTP or HTTPS URL (as opposed to
relative links inside the same website) are now passed to the linkcheck
builder so that it can do its checking work on them.

This necessitates requiring Sphinx >= 7.1 due to a restructuring of the
linkcheck builder code.

Fixes #3